refactor: accessToken 관련 로직 한 파일로 합침
